refactor(mf): timer - scss into css
