+++ micro-frontends/src/mf-timer/src/ui/modules/controller/controller.module.css
@@ -2,6 +2,6 @@
 	display: flex;
 	justify-content: center;
 	align-items: center;
-	column-gap: 15px;
-	margin-bottom: 15px;
+	column-gap: var(--ds-spacing-xl);
+	margin-bottom: var(--ds-spacing-xl);
 }
