+++ micro-frontends/src/mf-timer/src/ui/modules/player/player.tsx
@@ -1,5 +1,5 @@
 import type { FunctionComponent } from "react"
-import styles from "./player.module.scss"
+import styles from "./player.module.css"
 import { Reset } from "./reset/reset"
 import { StartAndStop } from "./start-and-stop/start-and-stop"
