+++ micro-frontends/src/mf-timer/src/ui/modules/presets/presets.module.css
@@ -2,8 +2,8 @@
 	display: flex;
 	justify-content: center;
 	align-items: center;
-	gap: 8px;
-	grid-area: 1/1/2/7;
+	gap: var(--ds-spacing-sm);
+	grid-area: 1 / 1 / 2 / 7;
 	margin-right: 0;
 	margin-left: 0;
 }
