+++ micro-frontends/src/mf-timer/src/ui/modules/screen/screen.module.css
@@ -1,0 +1,22 @@
+.screen {
+	display: flex;
+	grid-area: 3 / 1 / 4 / 7;
+	justify-self: center;
+	background-color: var(--ds-background-color-secondary);
+	max-width: 203px;
+	width: 100%;
+	height: 48px;
+	padding: var(--ds-spacing-md);
+	box-shadow: var(--ds-border-field);
+	font-family: var(--ds-font-sans-serif-bold);
+	font-size: var(--ds-font-size-xl);
+	line-height: var(--ds-line-height-normal);
+	letter-spacing: 3px;
+	color: var(--ds-text-color-primary);
+	user-select: none;
+	margin: 0 auto 15px;
+}
+
+.time {
+	width: 55px;
+}
