--- micro-frontends/src/mf-timer/src/ui/modules/screen/screen.module.scss
@@ -1,20 +1,0 @@
-.screen {
-  display: flex;
-  grid-area: 3/1/4/7;
-  justify-self: center;
-  background-color: var(--background-color-white);
-  max-width: 203px;
-  width: 100%;
-  height: 48px;
-  padding: 5px;
-  box-shadow: var(--border-field);
-  @include font('bold', 39, 44);
-  letter-spacing: 3px;
-  color: var(--text-color);
-  user-select: none;
-  margin: 0 auto 15px;
-}
-
-.time {
-  width: 55px;
-}
