+++ micro-frontends/src/mf-timer/src/ui/modules/screen/screen.tsx
@@ -1,6 +1,6 @@
 import { observer } from "mobx-react-lite"
 import type { FunctionComponent } from "react"
-import styles from "./screen.module.scss"
+import styles from "./screen.module.css"
 import { useScreen } from "./use-screen.hook"
 
 export const Screen: FunctionComponent = observer(() => {
