--- micro-frontends/src/mf-timer/src/ui/modules/wrapper/wrapper.module.scss
@@ -1,6 +1,0 @@
-.wrapper {
-  background-color: var(--background-color);
-  padding: 0 10px 10px;
-  max-width: 385px;
-  width: 100%;
-}
